Show a 2D map of the track in the upper pane
For this the functionality to receive vectors of GPX point properties
has been moved to gpxalyzer. The `App` is now initialized with
the boundary of the GPX files which are defined to have a 25% border
of the distance of the map. This is visually rather pleasing.

Moving the map is deactived but should be enabled and zooming should
also be possible in the future. For now this is a slow approach that
zips the different properties into tuples. Iterating over the objects
could be faster.